apify-job-worker/requirements.txt
temporalio>=1.7.0
httpx>=0.25.0
asyncpg>=0.29.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pydantic-ai>=0.0.14
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
zep-cloud>=2.0.0
